+++ src/schemathesis/generation/__init__.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 import random
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from enum import Enum
-from typing import Union, Iterable
+from typing import Union, Iterable, TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from hypothesis.strategies import SearchStrategy
 
 
 class DataGenerationMethod(str, Enum):
@@ -59,6 +62,13 @@ def generate_random_case_id(length: int = 6) -> str:
 
 
 @dataclass
+class HeaderConfig:
+    """Configuration for generating headers."""
+
+    strategy: SearchStrategy[str] | None = None
+
+
+@dataclass
 class GenerationConfig:
     """Holds various configuration options relevant for data generation."""
 
@@ -66,3 +76,5 @@ class GenerationConfig:
     allow_x00: bool = True
     # Generate strings using the given codec
     codec: str | None = "utf-8"
+    # Header generation configuration
+    headers: HeaderConfig = field(default_factory=HeaderConfig)